+++ sickchill/views/__init__.py
@@ -21,13 +21,13 @@
 from __future__ import print_function, unicode_literals
 
 from calendar import CalendarHandler
-from config import (Config, ConfigAnime, ConfigBackupRestore, ConfigNotifications, ConfigPostProcessing, ConfigProviders, ConfigSearch, ConfigShares,
-                    ConfigSubtitles)
 
 from api import ApiCall, ApiHandler, function_mapper, KeyHandler
 from authentication import LoginHandler, LogoutHandler
 from changelog import HomeChangeLog
 from common import PageTemplate
+from config import (Config, ConfigAnime, ConfigBackupRestore, ConfigNotifications, ConfigPostProcessing, ConfigProviders, ConfigSearch, ConfigShares,
+                    ConfigSubtitles)
 from history import History
 from home import Home
 from index import BaseHandler, WebHandler, WebRoot
